AM-449 : Restructured JUnit test cases
