+++ src/test/java/uk/gov/hmcts/reform/ccd/document/am/controller/CaseDocumentAmControllerTest.java
@@ -1,10 +1,9 @@
-package uk.gov.hmcts.reform.ccd.document.am;
+package uk.gov.hmcts.reform.ccd.document.am.controller;
 
 import org.junit.Before;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import uk.gov.hmcts.reform.ccd.document.am.controller.CaseDocumentAmController;
 import uk.gov.hmcts.reform.ccd.document.am.model.StoredDocumentHalResource;
 
 import java.util.UUID;
